@@ -1,2 +1,2 @@
 from .__version__ import __title__, __description__, __version__  # noqa
-from . import common, s3, glue, athena  # noqa
+from . import common, s3, glue, athena, redshift  # noqa
