@@ -2,5 +2,5 @@ __title__ = "awswrangler"
 __description__ = (
     "The missing Link between AWS services and the most popular Python data libraries"
 )
-__version__ = "0.0b2"
+__version__ = "0.0b3"
 __license__ = "Apache License 2.0"
